authentication & authorisation

+++ routers/items.py
@@ -46,18 +46,44 @@ def get_item_by_id(id: int, db: Session=Depends(get_db)):
 
 @router.put("/update/{id}", tags=["items"])
 def update_item_by_id(id: int, item: ItemCreate, db: Session=Depends(get_db), token:str=Depends(oauth2_scheme)):
+    try:
+        payload = jwt.decode(token, setting.SECRET_KEY, algorithms=setting.ALGORITHM)
+        username = payload.get("sub")
+        if username is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="unable to verify credential")
+        user = db.query(User).filter(User.email==username).first()
+        if user is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="unable to verify credential")
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="unable to verify credential")
     existing_item = db.query(Items).filter(Items.id==id)
     if not existing_item.first():
         return {"message": f"no details exist for Item ID {id}"}
-    existing_item.update(jsonable_encoder(item))
-    db.commit()
-    return {"message": f"Detail for item ID {id} successsfully updated"}
+    if existing_item.first().owner_id == user.id:
+        existing_item.update(jsonable_encoder(item))
+        db.commit()
+        return {"message": f"Detail for item ID {id} successsfully updated"}
+    else:
+        return {"meassage": {"you are not authorized"}}
 
 @router.delete("/item/delete/{id}", tags=["items"])
 def delete_item_by_id(id:int, db: Session=Depends(get_db), token:str=Depends(oauth2_scheme)):
+    try:
+        payload = jwt.decode(token, setting.SECRET_KEY, algorithms=setting.ALGORITHM)
+        username = payload.get("sub")
+        if username is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="uable to verify credential")
+        user = db.query(User).filter(User.email==username).first()
+        if user is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="unable to verify credential")
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="unable to verify credential")
     existing_item = db.query(Items).filter(Items.id==id)
     if not existing_item.first():
         return {"message" f"No detail exist for item ID {id}"}
-    existing_item.delete()
-    db.commit()
-    return {"message": f"Item with ID {id} is deleted"}    
+    if existing_item.first().owner_id == user.id:
+        existing_item.delete()
+        db.commit()
+        return {"message": f"Item with ID {id} is deleted"} 
+    else:
+        return {"message": "you are not authorized"}
